+++ Bricks1/State/State.swift
@@ -20,7 +20,7 @@ struct AppState: StateType {
     var streak: Int = 0
     var rank: Int = 0
     var totalUsers: Int = 0
-    var consistency: Double = 0.0
+    var consistency: Int = 0
     var countGradedTasks: Int = 0
     
     var loggedIn: Bool = false
